+++ kirksdk/qcos_client.go
@@ -881,14 +881,18 @@ func (p *qcosClientImp) GetContainerLogsRealtime(ctx context.Context, ip, since,
 	if !isSSL {
 		conn, err = net.DialTimeout("tcp", host, 10*time.Second)
 		if err != nil {
-			conn.Close()
+			if conn != nil {
+				conn.Close()
+			}
 			err = fmt.Errorf("net.DialTimeout err: %v", err)
 			return
 		}
 	} else {
 		conn, err = tls.DialWithDialer(&net.Dialer{Timeout: 10 * time.Second}, "tcp", host, nil)
 		if err != nil {
-			conn.Close()
+			if conn != nil {
+				conn.Close()
+			}
 			err = fmt.Errorf("tls.DialWithDialer err: %v", err)
 			return
 		}