+++ kirksdk/kirksdk.go
@@ -1,3 +1,3 @@
 package kirksdk
 
-const Version = "2.4.0"
+const Version = "3.0.0"