+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # vNext
+- 修复GetContainerLogsRealtime连接错误导致panic的bug
 
 # Release 2.4.0
 - 新增app配额查询接口