+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # vNext
 - 修复GetContainerLogsRealtime连接错误导致panic的bug
+- 下线CreateServiceVolume/SyncCreateServiceVolume接口
 
 # Release 2.4.0
 - 新增app配额查询接口