+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # vNext
+- 修复GetContainerLogsRealtime连接错误导致panic的bug
 - 下线CreateServiceVolume/SyncCreateServiceVolume接口
 
 # Release 2.4.0