+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # vNext
+
+# Release 3.0.0
 - AP详情添加enabled字段，表示这个AP是否已被管理员禁用
 - 修复GetContainerLogsRealtime连接错误导致panic的bug
 - 下线CreateServiceVolume/SyncCreateServiceVolume接口